+++ package.json
@@ -15,5 +15,10 @@
     "react-dom": "17.0.2",
     "remark": "^13.0.0",
     "remark-html": "^13.0.2"
+  },
+  "devDependencies": {
+    "@types/node": "^16.10.1",
+    "@types/react": "^17.0.24",
+    "typescript": "^4.4.3"
   }
 }
